Creating client and server folders for full stack project
