+++ client/package.json
@@ -1,13 +1,12 @@
 {
-  "name": "weather-app",
+  "name": "client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "express": "^4.18.2",
-    "pg": "^8.8.0",
+    "bootstrap": "^5.3.0-alpha1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
