+++ client/src/components/EditList.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const EditList = () =>{
+    return(
+        <h1>List</h1>
+    )
+};
+
+export default EditList;
