+++ client/src/components/ListWeather.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const ListWeather = () =>{
+    return(
+        <h1>List</h1>
+    )
+};
+
+export default ListWeather;
