+++ client/src/components/SearchWeather.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const SearchWeather = () =>{
+    return(
+        <h1>List</h1>
+    )
+};
+
+export default SearchWeather;
