--- weather-app/README.md
@@ -1,25 +1,0 @@
-# Weather App (name tbd)
-
-## *Authors*
-    - Davion Roachford 
-    - Aidan Papineau 
-    - Aaron Thomas 
-    - Allison Lowe
-
-## App Summary
-  -  Weather app will initially show users current location with current city name, temperature, high and low, and weather condition. The app will also allow users to add weather for locations around the world.  The user may also click on a given location and be provided with a more in depth weather report inclduing, 7 day forecast, hourly forecast, humidity, precipitation, and more.
-
-## Formating
-| **Front-End** | **Back-End** |
-| --- | --- |
-| React Redux | PostgresSQL |   
-
-## Task List
-- [ ] 
-- [ ]
-- [ ]
-
-## API 
-[API Reference](https://www.visualcrossing.com/resources/documentation/weather-api/timeline-weather-api/)
-
-[API Key](HSYWPZH9LXRTYUNAC5HYXZLX6)
