--- weather-app/src/.env
@@ -1,1 +1,0 @@
-api_key=
